+++ src/entity/PlaylistPodcastProgram.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, Index, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { Playlist } from './Playlist';
-import { PodcastProgram } from './PodcastProgram';
+import { PodcastProgram } from '~/podcast/entity/PodcastProgram';
 
 @Index('playlist_podcast_program_pkey', ['id'], { unique: true })
 @Entity('playlist_podcast_program', { schema: 'public' })
