+++ src/entity/PodcastProgramRole.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, Index, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { Member } from '~/member/entity/member.entity';
-import { PodcastProgram } from './PodcastProgram';
+import { PodcastProgram } from '~/podcast/entity/PodcastProgram';
 
 @Index('podcast_program_role_pkey', ['id'], { unique: true })
 @Entity('podcast_program_role', { schema: 'public' })
