+++ src/entity/index.ts
@@ -112,9 +112,9 @@ import { PlaylistPodcastProgram } from './PlaylistPodcastProgram';
 import { Podcast } from './Podcast';
 import { PodcastAlbum } from '~/podcast/entity/PodcastAlbum';
 import { PodcastAlbumCategory } from './PodcastAlbumCategory';
-import { PodcastAlbumPodcastProgram } from './PodcastAlbumPodcastProgram';
+import { PodcastAlbumPodcastProgram } from '~/podcast/entity/PodcastAlbumPodcastProgram';
 import { PodcastPlan } from './PodcastPlan';
-import { PodcastProgram } from './PodcastProgram';
+import { PodcastProgram } from '~/podcast/entity/PodcastProgram';
 import { PodcastProgramAudio } from './PodcastProgramAudio';
 import { PodcastProgramBody } from './PodcastProgramBody';
 import { PodcastProgramCategory } from './PodcastProgramCategory';
